requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
pypdf>=3.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
easyocr>=1.7.0
Pillow>=10.0.0
opencv-python>=4.8.0
openai>=1.0.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
